+++ code/packages/create-frontend-app/template/+nextjs+websockets/new-types.d.ts
@@ -1,8 +1,0 @@
-interface Window {
-  weave: Weave;
-  weaveTextEditing?: boolean;
-  weaveDragImageURL?: string;
-  colorTokenDragColor?: string;
-}
-
-declare module 'platform-detect';